Made campaign changes on WebApp, and this code is keeping the repositories synchronized. Moved CampaignStartIntro into common/pages folder so we can introduce the process on either Campaigns or WebApp. Created parallel js/components/Style/pageLayoutStyles.js to mimic WebApp, but we don't want this file to exactly match what is in WebApp because of all of the Cordova-specific code in this WebApp file. Switched from "theme.boxStyles.default" to "standardBoxShadow".
