+++ src/js/components/Style/pageLayoutStyles.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+// This is needed for pages in js/common/pages
+// eslint-disable-next-line import/prefer-default-export
+export const PageContentContainer = styled('div')`
+`;
